rest config + association stuff

+++ src/main/java/Service/IAssociation.java
@@ -2,13 +2,12 @@ package Service;
 
 import Entities.Association;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public interface IAssociation {
     List<String> getAssociationsName();
     Long deleteAssociation(Long id);
-    Long addAssociation(String data);
+    int addAssociation(String data);
     Long updateAssociation(String data);
     Association getAssociationById(Long id);
 }
